feat: 종목검색 api 구현

See merge request s11-bigdata-dist-sub1/S11P21A607!36

+++ backend/spring/src/main/java/com/example/mutualrisk/asset/entity/AssetHistory.java
@@ -3,6 +3,7 @@ package com.example.mutualrisk.asset.entity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Table(name = "asset_history")
@@ -24,5 +25,5 @@ public class AssetHistory {
     private LocalDateTime date;
 
     @Column(name = "price", nullable = false)
-    private Integer price;
+    private Double price;
 }
